fix(providers): a streamed tool call keeps its id and its arguments
A tool call streamed from Anthropic arrived as two calls, and the second
one had no id. `content_block_start` numbered the call, handed the
collector its id and its name, and then advanced the counter; every
`input_json_delta` after it read the counter it had already moved, so the
call's own arguments were filed one index further on. The collector keys
by index, so it built one call carrying an id and no arguments and a
second carrying arguments, an empty id and an empty name.

The empty id went into the conversation, and the next turn was rejected
before the model saw it:

    HTTP 400 messages.2.content.1.tool_use.id:
    String should match pattern '^[a-zA-Z0-9_-]+$'

An empty string does not match a pattern that wants at least one
character. That killed every streamed run that called a tool on the
native Anthropic provider, which is any run of a bundled agent whose
stage names a Claude model and whose machine has an Anthropic key -
measured with the bundled researcher on `--task "say hi"`, which now
finishes where it used to fail on its third turn.

The counter is now the index of the block that is *open* rather than the
number to hand out next, and both events prefer the `index` Anthropic
puts on every event over any counting of our own. The fallback is only
for a stream that numbers nothing, and it still keeps two calls apart.
